drv: register driver as filesystem
Because we called IoRegisterFileSystem() the system will no longer
call our DriverUnload() method.

The documentation explicitly says that a file system driver should not
be unloaded, but it is easier to test that everything works alright by
loading and unloading the driver a couple of times.

This function will only be called by the system when it will determine
that the device object has no more references. To do this we must use
an external mechanism to signal the driver to decrement these
references.

The solution is to use a private IOCTL:

- create a Win32 application that opens the DosDevice symbolic link we
  created (in NTK it's name is \DosDevices\lklvfs, but in Win32 it's
  \\.\lklvfs). This application sends a private IOCTL to the opened
  device and closes its handle.

- the driver receives the IOCTL and prepares itself to unload:

  - it must unregister itself as a file system driver
    (IoUnregisterFileSystem) and must delete the device object;

  - it must cleanup all open data and wait until all its resources can
    be safely deleted;

  - the driver's UnloadDriver function can now be called to properly
    unload the driver. If this function is called before the file
    system data and resources have been released, it can wait on some
    synchronization variables that will only be triggered when the
    driver decides it's ready to unload.

To implement these the driver needs to be able to respond to:
- IRP_MJ_CREATE - called when the device object
  is opened from user space;
- IRP_MJ_DEVICE_CONTROL - to be able to receive
  IOCTLs from user space;
- IRP_MJ_CLEANUP - would be called to close the
  handle, but this is optional.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>